Command changes and reorganization
