--- crosshatch/help.py
@@ -1,3 +1,0 @@
-HELP_MARKDOWN = '''# Help
-test text go brr
-'''
